third_party/blink/web_tests/svg/dynamic-updates/SVGFEOffsetElement-dom-dy-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dy' attribute of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS offsetElement.getAttribute('dy') is "50"
PASS offsetElement.getAttribute('dy') is "0"
PASS successfullyParsed is true

TEST COMPLETE
